deber 1

+++ src/main/java/com/distribuida/repo/BookRepository.java
@@ -5,5 +5,5 @@ import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
-public class BookRepository implements PanacheRepositoryBase<Book,Long> {
+public class BookRepository implements PanacheRepositoryBase<Book,Integer> {
 }
